Refactor EFCore tests
